teak generate aero BUG model
