+++ examples/BUG/NASTRAN/base_model.bdf
@@ -1,0 +1,13 @@
+$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
+$ MODEL										  $
+$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$
+
+INCLUDE './Parts/BUG_Fuselage_VTP.bdf'
+INCLUDE './Parts/BUG_WING_LWBOX.bdf'
+INCLUDE './Parts/BUG_WING_RWBOX.bdf'
+INCLUDE './Parts/fuselage_LWBOX_rbe.bdf'
+INCLUDE './Parts/fuselage_RWBOX_rbe.bdf'
+INCLUDE './Parts/MTOW_FUEL_LWBOX.bdf'
+INCLUDE './Parts/MTOW_FUEL_RWBOXmod.bdf'
+INCLUDE './Parts/BUG_LHTP.bdf'
+INCLUDE './Parts/BUG_RHTP.bdf'
